refactor: organize agent modules into packages
