+++ backend/src/agent/__init__.py
@@ -1,3 +1,7 @@
-from agent.graph_v2 import graph
+"""Agent package initialization."""
 
-__all__ = ["graph"]
+try:
+    from .graph import graph
+    __all__ = ["graph"]
+except Exception:  # pragma: no cover - optional graph import
+    __all__ = []
