+++ backend/src/agent/api/app.py
@@ -15,13 +15,13 @@ import asyncio
 from contextlib import asynccontextmanager
 
 # Import our graph and configuration functions
-from agent.graph_v2 import (
+from agent.graph.graph_v2 import (
     graph
 )
-from agent.state_V2 import OverallState
-from agent.graph_wrapper import create_tracked_executor
-from agent.node_progress import cleanup_progress
-from agent.streaming_api import JobProgressStreamer
+from agent.graph.state_V2 import OverallState
+from agent.tracing.graph_wrapper import create_tracked_executor
+from agent.tracing.node_progress import cleanup_progress
+from agent.api.streaming_api import JobProgressStreamer
 from agent.tracing import configure_tracing, get_tracer, add_span_attribute, add_span_event
 
 # Request/Response models
