+++ backend/src/agent/configuration/__init__.py
@@ -1,0 +1,3 @@
+from .configuration import Configuration
+
+__all__ = ["Configuration"]
