+++ backend/src/agent/graph/__init__.py
@@ -1,0 +1,3 @@
+from .graph_v2 import graph
+
+__all__ = ["graph"]
