+++ backend/src/agent/graph/research_agent_graph.py
@@ -22,10 +22,10 @@ import json
 
 
 from langchain.globals import set_debug, set_verbose
-from agent.state_V2 import ProductSimple
-from agent.basic_tools import llm_with_tools, route_tools_by_messages, tools, BasicToolNode, tavily
-from agent.llm_setup import get_llm
-from agent.search_pattern import retry_llm_tool_call
+from agent.graph.state_V2 import ProductSimple
+from agent.utils.basic_tools import llm_with_tools, route_tools_by_messages, tools, BasicToolNode, tavily
+from agent.configuration.llm_setup import get_llm
+from agent.graph.search_pattern import retry_llm_tool_call
 from typing import Optional, List, Dict
 from pydantic import BaseModel, Field
