+++ backend/src/agent/tracing/__init__.py
@@ -1,0 +1,3 @@
+from .tracing import configure_tracing, get_tracer, add_span_attribute, add_span_event
+
+__all__ = ["configure_tracing", "get_tracer", "add_span_attribute", "add_span_event"]
