+++ backend/src/agent/tracing/node_progress.py
