+++ backend/src/agent/utils/citation_utils.py
